requirements.txt
# Server
torch==2.8.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.117.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pydantic==2.11.9
pydantic_core==2.33.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.56.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
transformers>=4.35
safetensors>=0.4.2
sentencepiece>=0.1.96


# Model (NL→SQL)
transformers>=4.35
safetensors
sentencepiece
torch     # macOS wheel includes MPS; CUDA users replace with a CUDA-index install

# Client
streamlit
requests
pandas
python-dotenv
